Cleanup pass 2 - everything is fixed now?
