+++ MinishCapRandomizerUI/Program.cs
@@ -11,6 +11,6 @@ internal static class Program
         // To customize application configuration such as set high DPI settings or default font,
         // see https://aka.ms/applicationconfiguration.
         ApplicationConfiguration.Initialize();
-        Application.Run(new UI.MinishCapRandomizerUI());
+        Application.Run(new UI.MainWindow.MinishCapRandomizerUI());
     }
 }
