+++ MinishCapRandomizerUI/UI/About/About.resx
