+++ MinishCapRandomizerUI/UI/Config/SettingPresets.cs
@@ -2,12 +2,12 @@
 {
 	public class SettingPresets
 	{
-		public List<string> SettingsPresets { get; set; } = new List<string>();
+		public List<string> SettingsPresets { get; set; } = new();
 
-		public List<string> CosmeticsPresets { get; set; } = new List<string>();
+		public List<string> CosmeticsPresets { get; set; } = new();
 
-		public List<string> SettingsWeights { get; set; } = new List<string>();
+		public List<string> SettingsWeights { get; set; } = new();
 
-		public List<string> CosmeticsWeights { get; set; } = new List<string>();
+		public List<string> CosmeticsWeights { get; set; } = new();
 	}
 }
