+++ MinishCapRandomizerUI/UI/MainWindow/MinishCapRandomizerUI.resx
