+++ RandomizerCore/Randomizer/Helpers/Models/Transition.cs
@@ -1,4 +1,4 @@
-﻿namespace RandomizerCore.Randomizer.Models;
+﻿namespace RandomizerCore.Randomizer.Helpers.Models;
 
 internal class Transition : ITransition
 {
